Move username functions to db file
